Moved binary_cross_entropy_with_logits to experimental as it was merged as an experimental PR (#7730)
